+++ CHANGELOG.md
@@ -12,6 +12,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 - Updated Catacomb Maps & Scripts.
 
+- Added Slayer Shop & Updated Slayer Shop Items.
+
+- Added New Quest "The Dark Totem".
+
+- Skotizo Encounter.
+
 ## [1.5.1] - 11-01-2024
 
 ### Added